INTDEV-1090 module system: split SourceLayer by protocol
Replace the monolithic DefaultSourceLayer with a Strategy pattern: one
leaf class per protocol (FileSourceLayer, GitSourceLayer) plus a
CompositeSourceLayer that dispatches on location scheme. The public
interface and factory (SourceLayer, FetchTarget, createSourceLayer) are
unchanged — this is purely an internal reshape that makes adding new
source types (e.g. a Cyberismo Module Hub) a matter of writing one new
leaf class and registering it, without touching existing ones.

Collapses redundant protocol branches in resolver.ts and check-updates.ts
that became unnecessary once FileSourceLayer's listRemoteVersions /
queryRemote return sensible defaults for sources that don't support
versioning.
