+++ tools/data-handler/test/modules/resolver.test.ts
@@ -319,8 +319,14 @@ describe('modules/resolver', () => {
     expect(source.listLog).not.toContain('https://example.com/B.git');
   });
 
-  it('file-source root: no listRemoteVersions call, ref/version undefined', async () => {
-    // Stage the file source on disk so fetch can read its cardsConfig.
+  it('file-source root with version range: ref/version remain undefined', async () => {
+    // The resolver delegates to `SourceLayer.listRemoteVersions`; file-source
+    // leaves return `[]` which the resolver treats as "source doesn't support
+    // versioning", leaving ref and version undefined. This test used to assert
+    // the resolver wouldn't even call `listRemoteVersions` for a file source,
+    // but that branch moved into the source layer itself — the resolver now
+    // uniformly asks and the leaf decides. The observable outcome (ref /
+    // version undefined for a file source with a range) is preserved.
     const staged = join(tempDir, 'file-src');
     await mkdir(join(staged, '.cards', 'local'), { recursive: true });
     await writeFile(
@@ -342,7 +348,6 @@ describe('modules/resolver', () => {
     expect(out).toHaveLength(1);
     expect(out[0].ref).toBeUndefined();
     expect(out[0].version).toBeUndefined();
-    expect(source.listLog).toEqual([]);
   });
 
   it('throws when no remote version satisfies the declared range', async () => {
@@ -372,9 +377,7 @@ describe('modules/resolver', () => {
           {
             cardKeyPrefix: 'A',
             name: 'A',
-            modules: [
-              { name: 'B', location: 'https://example.com/B1.git' },
-            ],
+            modules: [{ name: 'B', location: 'https://example.com/B1.git' }],
           },
         ],
         [
